change dtype of nfevs and nfevs_after_it to int in ExcitationSolveScipy

+++ excitationsolve/excitation_solve_scipy.py
@@ -193,8 +193,8 @@ class ExcitationSolveScipy:
 
         self.energies = np.array(self.energies)
         self.energies_after_it = np.array(self.energies_after_it)
-        self.nfevs = np.array(self.nfevs)
-        self.nfevs_after_it = np.array(self.nfevs_after_it)
+        self.nfevs = np.array(self.nfevs, dtype=int)
+        self.nfevs_after_it = np.array(self.nfevs_after_it, dtype=int)
         self.params = np.array(self.params)
 
         return result
